fix(#991 PR-B): split deployments.rs test module to fix CI file-size gate
CI's Coverage job failed: src/deployments.rs was already sitting exactly
at its grandfathered anti-monolith ceiling (2806 LOC, zero slack), and
PR-B's additions pushed it to 2882.

The file already uses the sibling-submodule split pattern for
`review_repro_deployments_health_teams` (declared `mod
review_repro_deployments_health_teams;`, resolved to
src/deployments/review_repro_deployments_health_teams.rs — no mod.rs
conversion needed, since a non-mod.rs file's submodules resolve to a
same-named subdirectory). Applied the exact same pattern to the inline
`mod tests { ... }` block (2023 lines, over 2x the production code):
extracted verbatim into src/deployments/tests.rs, declared as `mod
tests;` in deployments.rs. `cargo fmt` re-indents the extracted body
safely (unlike a naive line-based dedent, which corrupted YAML fixtures
inside raw string literals in an earlier attempt at this — a raw
string's content isn't Rust syntax rustfmt would touch, but it also isn't
something a blind `sed 's/^    //'` should touch either).

One fix inside the moved file: `include_str!("deployments.rs")` (a
structural source-scan test reading its own file) had to become
`include_str!("../deployments.rs")` — the relative path is resolved from
the including file's new location.

Net effect: src/deployments.rs drops to 858 LOC (well under both its old
2806 ceiling and the repo-wide 2500 MAX_LOC), so removed it from
tests/src_file_size_invariant.rs's GRANDFATHERED list per that test's own
instruction ("remove the entry once it drops under MAX_LOC so the main
scan re-arms for it").

Verified: cargo test --workspace --tests --features tray — identical
257-failure set vs the established baseline (pre-existing, unrelated
resource-contention flakes); all 53 deployments:: tests green;
src_file_size_invariant green. cargo fmt --check and cargo clippy
--all-targets --features tray -- -D warnings clean.

Refs #991

Agend-Agent: gapfix-dev3
Agend-Task: t-20260704035112409879-67777-0
Agend-Branch: feat/991-topic-binding-template
